- Separate parsing functionality into it's own module\n- Add 'untested' keyword\n- Include untested caveat
